put in a control systems test
found a bunch of issues in SimplePID.
Might make the D part of SimplePID Actually use the Derivative class.
going to work on the autonomous part now, I'll see if it makes sense to
use the existing command based system for the commands.
Also basic and generic drive commands so future students will be able to
get a robot moving in autonomous much easier. yay.