New autonomous base plan implemented
The plan: have a box for methods which would allow them to have a set
time out and interrupt.

This probably could be done much more efficiently, but this is what I
know how to do with my limited knowledge in threading.

I have a really long winded explanation about why I chose this method if
anyone wants any justification